+++ acceptance-tests/src/testFixtures/java/tech/pegasys/teku/test/acceptance/dsl/TekuBeaconNode.java
@@ -454,6 +454,16 @@ public class TekuBeaconNode extends TekuNode {
         MINUTES);
   }
 
+  public void waitForFinalizationAfter(final UInt64 newFinalizedEpoch) {
+    LOG.debug("Wait for finalized block");
+    waitFor(
+        () ->
+            assertThat(fetchStateFinalityCheckpointEpoch().orElseThrow())
+                .isGreaterThanOrEqualTo(newFinalizedEpoch),
+        2,
+        MINUTES);
+  }
+
   public void waitForMilestone(final SpecMilestone expectedMilestone) {
     waitForLogMessageContaining("Activating network upgrade: " + expectedMilestone.name());
   }
@@ -582,6 +592,18 @@ public class TekuBeaconNode extends TekuNode {
     final Bytes32 root = Bytes32.fromHexString(jsonNode.get("data").get("root").asText());
     final boolean executionOptimistic = jsonNode.get("execution_optimistic").asBoolean();
     return Optional.of(Pair.of(root, executionOptimistic));
+  }
+
+  private Optional<UInt64> fetchStateFinalityCheckpointEpoch() throws IOException {
+    final String result =
+        httpClient.get(getRestApiUrl(), "/eth/v1/beacon/states/head/finality_checkpoints");
+    if (result.isEmpty()) {
+      return Optional.empty();
+    }
+    final JsonNode jsonNode = OBJECT_MAPPER.readTree(result);
+    final UInt64 finalizedEpoch =
+        UInt64.valueOf(jsonNode.get("data").get("finalized").get("epoch").asText());
+    return Optional.of(finalizedEpoch);
   }
 
   private Optional<Bytes32> fetchBeaconHeadRoot() throws IOException {
@@ -775,5 +797,14 @@ public class TekuBeaconNode extends TekuNode {
 
   public void expectElOnline() throws IOException {
     assertThat(getStatusElOffline()).isFalse();
+  }
+
+  public BeaconState fetchFinalizedState() throws IOException {
+    final Bytes beaconStateBytes =
+        httpClient.getAsBytes(
+            getRestApiUrl(),
+            "eth/v2/debug/beacon/states/finalized",
+            Map.of("Accept", "application/octet-stream"));
+    return spec.deserializeBeaconState(Bytes.wrap(beaconStateBytes));
   }
 }