+++ ethereum/statetransition/src/main/java/tech/pegasys/teku/statetransition/block/BlockImporter.java
@@ -182,9 +182,8 @@ public class BlockImporter {
     if (finalizedCheckpoint != null
         && recentChainData
             .getStore()
-            .getLatestFinalized()
-            .getRoot()
-            .equals(finalizedCheckpoint.getRoot())) {
+            .getFinalizedCheckpoint()
+            .equals(finalizedCheckpoint.getCheckpoint())) {
       return SafeFuture.completedFuture(finalizedCheckpoint);
     }